refactor: refactor according to review changes-requested
