+++ src/components/ui/__test__/Button.test.js
@@ -1,6 +1,6 @@
 import renderer from "react-test-renderer";
 
-import Button from "../buttons";
+import Button from "../Button";
 
 it("renders correctly", () => {
     const tree = renderer.create(<Button />).toJSON();
